fix: :card_file_box: reformatted points table migration on renaming columns to identify current state of db
In the case that columns are manually added to the DB (in the old format), the migration did not previously modify those columns. The migration naively adds a column to the database, creating two instances. One with the data (in the old format), and the copy (in the new format) with all zeroes. This change ensures that irrespective of the state of the database, the end state is in the new format with no data is lost, be it in the up or down migration.